Better name splitting
Normally, if a person's name entered into a peopleselect does not exist yet, the name will be split into its components ([title] [first_name] last_name) and a new user created. Now it is possible to enter a name in double quotes (") or with an underscore (_) in order to prevent splitting.